Making UpdateCommand

+++ src/Shopway.Application/Products/Commands/CreateProduct/CreateProductCommand.cs
@@ -5,7 +5,7 @@ namespace Shopway.Application.Products.Commands.CreateProduct;
 public sealed record CreateProductCommand
 (
     string ProductName,
-    int Price,
+    decimal Price,
     string UomCode,
     string Revision
 ) : ICommand<Guid>;
